requirements.txt
asyncio
python-telegram-bot 
yt-dlp
